fix bugs

+++ src/main/java/org/mesdag/particlestorm/mixed/ITextureAtlasSprite.java
@@ -1,7 +1,7 @@
 package org.mesdag.particlestorm.mixed;
 
 public interface ITextureAtlasSprite {
-    int particlestorm$getInvOx();
+    float particlestorm$getInvOx();
 
-    int particlestorm$getInvOy();
+    float particlestorm$getInvOy();
 }
